requirements.txt
-e .
aws-cdk.cdk
aws-cdk.aws-lambda
